requirements.txt
langchain==0.0.327
openai==0.28.0
pandas==1.2.4
pandas==2.0.3
python-dotenv==1.0.0
retrying==1.3.4
scipy==1.5.4
sentence_transformers==2.2.2
torch==1.11.0
